Fix #30 - Make OAuth2 token expires available

+++ Ramone/OAuth2/OAuth2AccessTokenResponse.cs
@@ -11,7 +11,7 @@ namespace Ramone.OAuth2
   {
     public string access_token { get; set; }
     public string token_type { get; set; }
-    public int? expires_in { get; set; }
+    public long? expires_in { get; set; }
     public string refresh_token { get; set; }
     public Hashtable AllParameters { get; set; }
   }
